replace world to maindisplay
