+++ src/components/MainDisplay.js
@@ -1,16 +1,92 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Grid from '@material-ui/core/Grid';
+import { makeStyles } from '@material-ui/core/styles';
+import Tabs from '@material-ui/core/Tabs';
+import Tab from '@material-ui/core/Tab';
+import Typography from '@material-ui/core/Typography';
+import Box from '@material-ui/core/Box';
 
 import MemoFamily from "./MemoFamily";
 
 export default function MainDisplay(props) {
-  return (
-    <Grid container component="main" className="left-container">
-      <CssBaseline/>
+  const classes = useStyles();
+  const [value, setValue] = React.useState(0);
 
-      <MemoFamily gridXs={12} gridMd={12} requiredMemoInput={true} displayClassName="MainDisplay-memos"/>
-    </Grid>
+  function handleChange(event, newValue) {
+    setValue(newValue);
+  }
+
+  const current_page = props.current_page;
+
+  return (
+    <div className={classes.root}>
+      <Tabs
+        orientation="vertical"
+        variant="scrollable"
+        value={value}
+        onChange={handleChange}
+        aria-label="Vertical tabs example"
+        className={classes.tabs}
+      >
+        / selected not working
+        <Tab selected={current_page == 'world'} label="World" {...a11yProps()} />
+        <Tab selected={current_page == 'home'} label="home" {...a11yProps(1)} />
+        <Tab selected={current_page == 'profile'} label="profile" {...a11yProps(2)} />
+      </Tabs>
+
+      <TabPanel value={value} index={0} className={classes.memos}>
+        <MemoFamily gridXs={12} gridMd={12} requiredMemoInput={true} displayClassName=""/>
+      </TabPanel>
+      <TabPanel value={value} index={1}>
+        coming soon...
+      </TabPanel>
+    </div>
   );
 }
+
+
+function TabPanel(props) {
+  const { children, value, index, ...other } = props;
+
+  return (
+    <Typography
+      component="div"
+      role="tabpanel"
+      hidden={value !== index}
+      id={`vertical-tabpanel-${index}`}
+      aria-labelledby={`vertical-tab-${index}`}
+      {...other}
+    >
+      <Box p={3}>{children}</Box>
+    </Typography>
+  );
+}
+
+TabPanel.propTypes = {
+  children: PropTypes.node,
+  index: PropTypes.any.isRequired,
+  value: PropTypes.any.isRequired,
+};
+
+function a11yProps(index) {
+  return {
+    id: `vertical-tab-${index}`,
+    'aria-controls': `vertical-tabpanel-${index}`,
+  };
+}
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    flexGrow: 1,
+    backgroundColor: theme.palette.background.paper,
+    display: 'flex',
+    height: 224,
+  },
+  tabs: {
+    borderRight: `1px solid ${theme.palette.divider}`,
+  },
+  memos: {
+    width: '50%',
+  },
+}));
