--- src/components/WorldDisplay.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Grid from '@material-ui/core/Grid'
-
-import Menu from "./Menu";
-
-export default function WorldDisplay(props) {
-  return (
-      <Menu/>
-  );
-}
